@@ -1,68 +1,173 @@
-<x-filament::layouts.base :title="$title">
-    <a href="#content" class="sr-only">Skip to content</a>
+<!DOCTYPE html>
+<html lang="{{ str_replace('_', '-', app()->getLocale()) }}" dir="{{ trans('filament::layout.direction') === 'rtl' ? 'rtl' : 'ltr' }}" class="antialiased bg-gray-100 js-focus-visible">
+    <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <meta name="csrf-token" content="{{ csrf_token() }}">
 
-    <div class="relative overflow-hidden"
-         role="group"
-         tabindex="-1"
-         x-data="{ headerIsOpen: false }"
-         @keydown.escape.window="headerIsOpen = false"
-         x-on:resize.window="if (window.outerWidth > 768) headerIsOpen = false"
-    >
-        <div class="fixed z-20 flex w-64 h-screen transition-transform duration-200 transform ltr:md:translate-x-0 rtl:md:translate-x-0"
-             :class="headerIsOpen ? 'translate-x-0' : 'ltr:-translate-x-full rtl:translate-x-full'">
-            <header role="banner"
-                    tabindex="-1"
-                    id="banner"
-                    class="flex flex-col flex-grow space-y-4 overflow-y-auto text-gray-400 bg-gray-800 shadow-lg md:shadow-none">
-                <x-filament::branding.app />
+        <title>{{ $title ? "{$title} - " : null }} {{ config('app.name') }}</title>
 
-                <x-filament-nav class="flex-grow px-4 overflow-y-auto" />
+        <style>[x-cloak] { display: none !important; }</style>
 
-                <x-filament::dropdown
-                    class="flex items-center flex-grow w-full p-4 space-x-3 rtl:space-x-reverse text-left transition-colors duration-200 hover:text-white hover:bg-gray-900">
-                    <x-slot name="button">
-                        <x-filament-avatar :user="\Filament\Filament::auth()->user()" :size="32" class="flex-shrink-0 w-8 h-8 rounded-full" />
+        @livewireStyles
 
-                        <span class="flex-grow text-sm font-medium leading-tight">{{ \Filament\Filament::auth()->user()->name }}</span>
-                    </x-slot>
+        <link rel="preconnect" href="https://fonts.googleapis.com">
+        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+        <link href="https://fonts.googleapis.com/css2?family=DM+Sans:ital,wght@0,400;0,500;0,700;1,400;1,500;1,700&family=JetBrains+Mono&display=swap" rel="stylesheet" />
 
-                    @if (\Filament\Filament::auth()->user()->isFilamentAdmin())
-                        <x-filament::dropdown-link href="{{ route('filament.users.index') }}">
-                            {{ __('filament::nav.dropdown.users.label') }}
-                        </x-filament::dropdown-link>
-                    @endif
+        <link rel="stylesheet" href="{{ route('filament.asset', [
+            'id' => \Filament\get_asset_id('app.css'),
+            'path' => 'app.css',
+        ]) }}" />
 
-                    <x-filament::dropdown-link href="{{ route('filament.account') }}">
-                        {{ __('filament::nav.dropdown.account.label') }}
-                    </x-filament::dropdown-link>
+        @foreach (\Filament\Facades\Filament::getStyles() as $path)
+            <link rel="stylesheet" href="{{ $path }}" />
+        @endforeach
+    </head>
 
-                    <livewire:filament.core.auth.logout class="w-full px-4 py-2 text-gray-600 transition-colors duration-200 hover:bg-gray-200" />
-                </x-filament::dropdown>
-            </header>
-
-            <button
-                type="button"
-                aria-controls="banner"
-                @click.prevent="headerIsOpen = false"
-                :aria-expanded="headerIsOpen"
+    <body>
+        <div class="flex min-h-screen w-full bg-gray-100">
+            <div
+                x-data="{}"
                 x-cloak
-                x-show.opacity="headerIsOpen"
-                class="absolute ltr:right-0 rtl:left-0 p-3 text-gray-200 transition-colors duration-200 transform ltr:translate-x-full rtl:-translate-x-full md:hidden top-2 hover:text-white">
-                <x-heroicon-o-x class="w-6 h-6" />
-            </button>
-        </div>
+                x-show="$store.sidebar.isOpen"
+                x-transition.opacity.500ms
+                x-on:click="$store.sidebar.close()"
+                class="fixed inset-0 z-20 w-full h-full bg-gray-900/50 lg:hidden"
+            ></div>
 
-        <span class="absolute inset-0 z-10 bg-gray-800 bg-opacity-50 md:hidden" x-cloak x-show="headerIsOpen"
-              @click="headerIsOpen = false"></span>
+            <aside
+                x-data="{}"
+                x-cloak
+                x-bind:class="$store.sidebar.isOpen ? 'translate-x-0' : '-translate-x-full'"
+                class="fixed inset-y-0 left-0 z-20 flex flex-col h-screen overflow-hidden shadow-2xl rounded-r-2xl transition duration-300 bg-white lg:border-r w-80 lg:z-0 lg:translate-x-0"
+            >
+                <header class="border-b h-[4rem] px-6 flex items-center">
+                    <a class="text-xl font-bold tracking-tight" href="{{ \Filament\Pages\Dashboard::geturl() }}">
+                        {{ config('app.name') }}
+                    </a>
+                </header>
 
-        <div class="flex flex-col w-full min-h-screen ltr:md:pl-64 rtl:md:pr-64">
-            <div class="flex-grow">
-                {{ $slot }}
+                <nav class="flex-1 py-6">
+                    <ul class="space-y-6 px-6">
+                        @foreach (\Filament\Facades\Filament::getNavigation() as $group => $items)
+                            <li>
+                                @if ($group)
+                                    <p class="font-bold uppercase text-gray-600 text-xs tracking-wider">
+                                        {{ $group }}
+                                    </p>
+                                @endif
+
+                                <ul @class([
+                                        'text-sm space-y-1 -mx-3',
+                                        'mt-2' => $group,
+                                    ])>
+                                    @foreach ($items as $item)
+                                        @php
+                                            $isActive = $item->isActive();
+                                        @endphp
+                                        <li>
+                                            <a
+                                                href="{{ $item->getUrl() }}"
+                                                @class([
+                                                    'flex items-center gap-3 px-3 py-2 rounded-lg font-medium transition',
+                                                    'hover:bg-gray-500/5 focus:bg-gray-500/5' => ! $isActive,
+                                                    'bg-primary-500 text-white' => $isActive,
+                                                ])
+                                            >
+                                                <x-dynamic-component :component="$item->getIcon()" class="h-5 w-5" />
+
+                                                <span>
+                                                        {{ $item->getLabel() }}
+                                                    </span>
+                                            </a>
+                                        </li>
+                                    @endforeach
+                                </ul>
+                            </li>
+
+                            @if (! $loop->last)
+                                <li>
+                                    <div class="border-t -mr-6"></div>
+                                </li>
+                            @endif
+                        @endforeach
+                    </ul>
+                </nav>
+
+                <footer class="border-t px-6 py-3 flex items-center gap-3">
+                    <div
+                        class="w-11 h-11 rounded-full bg-gray-200 bg-cover bg-center"
+                        style="background-image: url('{{ \Filament\Facades\Filament::getAvatar(auth()->user()) }}')"
+                    ></div>
+
+                    <div>
+                        <p class="text-sm font-bold">
+                            {{ auth()->user()->getFilamentName() }}
+                        </p>
+
+                        <p class="text-xs text-gray-500 transition hover:text-gray-700 focus:text-gray-700">
+                            <a href="{{ route('filament.logout') }}">
+                                Sign out
+                            </a>
+                        </p>
+                    </div>
+                </footer>
+            </aside>
+
+            <div class="w-screen flex-1 lg:pl-80">
+                <header class="h-[4rem] border-b flex items-center">
+                    <div class="flex items-center w-full max-w-6xl px-2 mx-auto sm:px-4 md:px-6 lg:px-8">
+                        <button x-data="{}" x-on:click="$store.sidebar.open()" class="flex-shrink-0 flex items-center justify-center w-10 h-10 text-primary-500 transition rounded-full hover:bg-gray-500/5 focus:bg-primary-500/10 focus:outline-none lg:hidden">
+                            <x-heroicon-o-menu class="w-6 h-6" />
+                        </button>
+
+                        <div class="flex-1 flex items-center justify-between">
+                            <div>
+                                <ul class="hidden gap-4 items-center font-medium text-sm lg:flex">
+                                    @foreach ($breadcrumbs as $url => $label)
+                                        <li>
+                                            <a
+                                                href="{{ is_int($url) ? '#' : $url }}"
+                                                @class([
+                                                    'text-gray-500' => $loop->last,
+                                                ])
+                                            >
+                                                {{ $label }}
+                                            </a>
+                                        </li>
+
+                                        @if (! $loop->last)
+                                            <li class="h-6 border-r border-gray-300 -skew-x-12"></li>
+                                        @endif
+                                    @endforeach
+                                </ul>
+                            </div>
+
+                            @livewire('filament.global-search')
+                        </div>
+                    </div>
+                </header>
+
+                <div class="max-w-6xl px-4 py-6 mx-auto md:px-6 lg:px-8">
+                    {{ $slot }}
+                </div>
             </div>
-
-            <footer rel="contentinfo" class="p-4 text-center md:px-6">
-                <x-filament::branding.footer />
-            </footer>
         </div>
-    </div>
-</x-filament::layouts.base>
+
+        @livewireScripts
+
+        <script>
+            window.filamentData = @json(\Filament\Facades\Filament::getScriptData());
+        </script>
+
+        <script src="{{ route('filament.asset', [
+            'id' => Filament\get_asset_id('app.js'),
+            'path' => 'app.js',
+        ]) }}"></script>
+
+        @foreach (\Filament\Facades\Filament::getScripts() as $path)
+            <script src="{{ $path }}"></script>
+        @endforeach
+    </body>
+</html>
