@@ -1,80 +1,35 @@
-<div>
-    <x-filament::app-header
-        :breadcrumbs="static::getBreadcrumbs()"
-        :title="$title"
-    >
-        <x-slot name="actions">
-            @if ($this->canDelete())
-                <x-filament::modal>
-                    <x-slot name="trigger">
-                        <x-filament::button
-                            x-on:click="open = true"
-                            color="danger"
-                        >
-                            {{ __(static::$deleteButtonLabel) }}
-                        </x-filament::button>
-                    </x-slot>
-
-                    <x-filament::card class="max-w-2xl space-y-5">
-                        <x-filament::card-header :title="static::$deleteModalHeading">
-                            <p class="text-sm text-gray-500">
-                                {{ __(static::$deleteModalDescription) }}
-                            </p>
-                        </x-filament::card-header>
-
-                        <div class="space-y-3 sm:space-y-0 sm:flex sm:space-x-3 rtl:space-x-reverse sm:justify-end">
-                            <x-filament::button
-                                x-on:click="open = false"
-                            >
-                                {{ __(static::$deleteModalCancelButtonLabel) }}
-                            </x-filament::button>
-
-                            <x-filament::button
-                                wire:click="delete"
-                                color="danger"
-                            >
-                                {{ __(static::$deleteModalConfirmButtonLabel) }}
-                            </x-filament::button>
-                        </div>
-                    </x-filament::card>
-                </x-filament::modal>
-            @endif
+<x-filament::page>
+    <x-filament::header :actions="$this->getActions()">
+        <x-slot name="heading">
+            {{ $title }}
         </x-slot>
-    </x-filament::app-header>
+    </x-filament::header>
 
-    <x-filament::app-content class="space-y-6">
-        @php
-            $schema = $this->getForm()->getSchema();
-        @endphp
+    <x-filament::form wire:submit.prevent="save">
+        {{ $this->form }}
 
-        <form wire:submit.prevent="save" class="space-y-6">
-            @if ($this->getForm()->hasWrapper())
-                <x-filament::card class="space-y-6">
-                    <x-forms::form :schema="$this->getForm()->getSchema()" :columns="$this->getForm()->getColumns()" />
+        <x-filament::actions :actions="$this->getFormActions()" />
+    </x-filament::form>
 
-                    <x-filament::resources.forms.actions :actions="$this->getActions()" />
-                </x-filament::card>
-            @else
-                <x-forms::form :schema="$this->getForm()->getSchema()" :columns="$this->getForm()->getColumns()" />
+    <x-filament::modal id="delete">
+        <x-slot name="heading">
+            Delete {{ $this->getRecordTitle() ?? static::getResource()::getLabel() }}
+        </x-slot>
 
-                <x-filament::resources.forms.actions :actions="$this->getActions()" />
-            @endif
-        </form>
+        <x-slot name="subheading">
+            Are you sure you would like to do this?
+        </x-slot>
 
-        <x-filament::resources.relations
-            :owner="$record"
-            :relations="static::getResource()::relations()"
-        />
-    </x-filament::app-content>
+        <x-slot name="actions">
+            <x-filament::modal.actions full-width>
+                <x-filament::button x-on:click="isOpen = false" color="secondary">
+                    Cancel
+                </x-filament::button>
 
-    <div
-        x-data
-        x-init="
-            Mousetrap.bindGlobal(['ctrl+s', 'command+s'], $event => {
-                $event.preventDefault()
-
-                document.querySelector(`button[type='submit']`).click()
-            })
-        "
-    ></div>
-</div>
+                <x-filament::button wire:click="delete" color="danger">
+                    Delete
+                </x-filament::button>
+            </x-filament::modal.actions>
+        </x-slot>
+    </x-filament::modal>
+</x-filament::page>
