@@ -6,5 +6,5 @@ use Filament\Models\Contracts\FilamentUser;
 
 interface AvatarProvider
 {
-    public function get(FilamentUser $user, $size = 48, $dpr = 1);
+    public function get(FilamentUser $user): string;
 }
